pyproject.toml
[project]
name = "tg-prognoz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.25.0",
    "aiosqlite>=0.22.1",
    "google-genai>=1.63.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
    "sqlalchemy>=2.0.46",
]
